parser: Split into multiple files, now parsing to struct file
